appengine/warmup: practical setup from main
